api/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
chromadb[fastembed]
sentence-transformers
numpy
